+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.0] - 2025-10-19
+- Windows compatibility
+- Change release strategy
+- Headers in ttypt folder
+
 ## [0.0.11] - 2025-10-02
 - System Logging
 - Debug Macros